llm-extracter/pyproject.toml
[tool.poetry]
name = "llm-extracter"
version = "0.1.0"
description = ""
authors = ["sumitskj <sumitjangirdss.1@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
openai = "^1.51.0"
tortoise-orm = "^0.21.7"
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
